Support YAML files for video metadata, sync/edit
This allows for initial support of editing YAML files for the metadata in
steve projects.  As fetching, uploading, steves webedit and push (still)
work with JSON files, this support relies on synchronisation of the
json and yaml directories.

Synchronisation works in both ways, but you cannot touch the metadata
for the same video on both the json and the yaml side. You *can* change
the metadata for different videos ( json/01.json and yaml/02.yaml) and
sync, which will update json/02.json and yaml/01.yaml respectively.

Syncing is done based on "last-synced" information kept in status.yaml
in the project directory. See the documentation for what to do if you
remove that file, or touched JSON and YAML metadata for the same video.

There are two new commandline subcommand:

- steve-cmd sync

  Synchronises the yaml and json directories, creating files if necessary.
  Should be invoked after pull, changing files with webedit and before
  push. Doesn't do anything if there are no changes, and first checks if
  there are any conflicts (files touched on both sided), before making
  any changes

- steve-cmd yaml

  This combines all YAML files into a single combined temporary YAML file
  and invokes your $EDITOR on it. The combined file contains information
  to split it up into the individual files again *only writing files that
  were actually changed*.
  If you want to use the combined file in another tool than your editor
  use:

     steve-cmd yaml --name combined.yaml --combine
  and:

     steve-cmd yaml --name combined.yaml --split